collector/requirements.txt
pymodbus==2.5.3
pyyaml==6.0
apscheduler==3.9.1
sqlalchemy==1.4.23
psycopg2-binary==2.9.1

